+++ backend/.eslintrc.js
@@ -32,7 +32,7 @@ module.exports = {
     ],
   },
   settings: {
-    'import/external-module-folders': ['../types'], // TODO : check
+    'import/external-module-folders': ['../types'],
     'import/resolver': {
       typescript: {},
       node: {
