+++ backend/src/friend/dto/requested-friend-response.dto.ts
@@ -1,4 +1,4 @@
-import { RequestedFriendsResponse } from '@/types/friend/response/requested-friend-response.interface';
+import { RequestedFriendsResponse } from '@/types/friend/response';
 
 import { UserInfoDto } from '../../user/dto/user-info.dto';
